+++ Libplanet/Blockchain/BlockChain.cs
@@ -551,6 +551,9 @@ namespace Libplanet.Blockchain
         /// are required for action execution and rendering.
         /// <see cref="StateCompleterSet{T}.Recalculate"/> by default.
         /// </param>
+        /// <exception cref="InvalidBlockBytesLengthException">Thrown when the block to mine is
+        /// too long (according to <see cref="IBlockPolicy{T}.GetMaxBlockBytes(long)"/>) in bytes.
+        /// </exception>
         /// <exception cref="InvalidBlockException">Thrown when the given
         /// <paramref name="block"/> is invalid, in itself or according to
         /// the <see cref="Policy"/>.</exception>
@@ -946,6 +949,15 @@ namespace Libplanet.Blockchain
             stateCompleters ??= StateCompleterSet<T>.Recalculate;
 
             _logger.Debug("Trying to append block {blockIndex}: {block}", block?.Index, block);
+
+            if (block.BytesLength > Policy.GetMaxBlockBytes(block.Index))
+            {
+                throw new InvalidBlockBytesLengthException(
+                    block.BytesLength,
+                    Policy.GetMaxBlockBytes(block.Index),
+                    "The block to append is too long in bytes."
+                );
+            }
 
             IReadOnlyList<ActionEvaluation> evaluations = null;
             _rwlock.EnterUpgradeableReadLock();